+++ examples/official-storybook/stories/addon-knobs/with-knobs-options.stories.js
@@ -10,7 +10,7 @@ export default {
   ],
 };
 
-export const acceptsOptions = () => <div>{text('Rendered string', '<h1>Hello</h1>')}</div>;
-acceptsOptions.story = {
+export const AcceptsOptions = () => <div>{text('Rendered string', '<h1>Hello</h1>')}</div>;
+AcceptsOptions.story = {
   name: 'accepts options',
 };